Move list API to separate cpp file
Fixes: #147
