+++ internal/c/libqb/build.mk
@@ -7,6 +7,7 @@ libqb-objs-y += $(PATH_LIBQB)/src/filesystem.o
 libqb-objs-y += $(PATH_LIBQB)/src/datetime.o
 libqb-objs-y += $(PATH_LIBQB)/src/error_handle.o
 libqb-objs-y += $(PATH_LIBQB)/src/gfs.o
+libqb-objs-y += $(PATH_LIBQB)/src/qblist.o
 libqb-objs-y += $(PATH_LIBQB)/src/mem.o
 libqb-objs-y += $(PATH_LIBQB)/src/rounding.o
 libqb-objs-y += $(PATH_LIBQB)/src/qbs.o
